requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
gunicorn>=20.0.0

# Database
psycopg2-binary>=2.9.0

# Face recognition (headless OpenCV - no X11/GTK deps)
opencv-python-headless>=4.5.0
Pillow>=9.0.0
numpy>=1.21.0

# Email verification (uses stdlib smtplib - no extra install needed)
# Required env vars for email:
#   SMTP_HOST=smtp.exmail.qq.com
#   SMTP_PORT=465  (or 587 for TLS)
#   SMTP_USER=Dean.hang@portraitpayai.com
#   SMTP_PASSWORD=<password>
